pyproject.toml
[project]
name = "collective-intelligence"
version = "0.1.0"
description = "A lightweight autonomous AI entity harness."
readme = "README.md"
requires-python = ">=3.13.2"
dependencies = [
    "anthropic>=0.96.0",
    "apscheduler>=3.10",
    "python-dotenv>=1.0.1",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "python-frontmatter>=1.1.0",
    "rich>=13.7",
    "google-api-python-client>=2.194.0",
    "google-auth-httplib2>=0.3.1",
    "google-auth-oauthlib>=1.3.1",
]

[tool.hatch.build.targets.wheel]
packages = ["harness"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
